outsource MetricWithArbitraryCounts again
